PR comments addressed - minor refactor and module move
